+++ src/physics.c
@@ -3,18 +3,18 @@
 #include <stdio.h>
 #include <string.h>
 
-double get_distance(Vector2 p1, Vector2 p2){
+float get_distance(Vector2 p1, Vector2 p2){
     return sqrt(pow(p1.x - p2.x, 2) + pow(p1.y - p2.y, 2));
 }
 static Vector2 get_gforce(Body b1, Body b2){
-    double d = get_distance(b1.position, b2.position) * SCALE_FACTOR;
-    double F = (G_CONST * b1.mass * b2.mass) / pow(d, 2);
+    float d = get_distance(b1.position, b2.position) * SCALE_FACTOR;
+    float F = (G_CONST * b1.mass * b2.mass) / pow(d, 2);
     Vector2 vF;
     vF.x = (F * (b2.position.x - b1.position.x)) / d;
     vF.y = (F * (b2.position.y - b1.position.y)) / d;
     return vF;
 }
-static void update_body_vel(Body* update, Body reference, double tSpeed){
+static void update_body_vel(Body* update, Body reference, float tSpeed){
     Vector2 vF = get_gforce(*update, reference);
     update->velocity.x += (vF.x / update->mass) * GetFrameTime() * tSpeed;
     update->velocity.y += (vF.y / update->mass) * GetFrameTime() * tSpeed;
@@ -34,7 +34,7 @@ static Body get_more_massive(Body b1, Body b2){
 static Body get_less_massive(Body b1, Body b2){
     return b1.mass < b2.mass ? b1 : b2;
 }
-static double get_new_radius(double r1, double r2){
+static float get_new_radius(float r1, float r2){
     return cbrt(pow(r1, 3) + pow(r2, 3));
 }
 
@@ -73,7 +73,7 @@ void update_collisions(Simulation* simulation){
         }
     }
 }
-void update_simulation(Simulation* simulation, double tSpeed){
+void update_simulation(Simulation* simulation, float tSpeed){
     for(unsigned i = 0; i < simulation->count; i++){
         for(unsigned j = i + 1; j < simulation->count; j++){
             if(i != j){
